moved NITI-only methods from BaseTextInputConnection to the NativeTextInputConnection, cleaned up some code, introduced NativeTextInputDelegate
